% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_utils.R
\name{capture_args}
\alias{capture_args}
\title{\code{capture_args}}
\usage{
capture_args()
}
\value{
a list of arguments passed
}
\description{
returns a list of all arguments passed to a function, such that you could re-call the function using \code{do.call()}
}
\details{
If you pass in a \code{data.frame}, such as \code{data_in}, the whole thing is copied in the respective argument, which is returned. If you have an object in memory called \code{data_i}, defined as \code{data.frame(a=1:3,b=letters[1:3])} and you do \code{f(data_in = data_i)}, then the output is not \code{list()}
}
\examples{
f <- function(x,y,...){
  capture_args()
}

g <- function(x=1,y){
  capture_args()
}
f(x=1,y=1:10)
f(x=1,y=1:10,z="constant",a="b")
g(y=1:10)

}
