% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R, R/lifetables.R
\name{check_data}
\alias{check_data}
\title{\code{check_data}}
\usage{
check_data(data_out)

check_data(data_out)
}
\arguments{
\item{data_out}{a data.frame or tibble. The data.frame output of the \code{lt_flexible} function.}

\item{data}{tibble. A tibble generated by the output of the read_data function.}
}
\value{
A data.frame with 3 columns and results of 9 tests: message - \code{check} - the name of the test, \code{message} - The error message with corresponding information generated if the test is failed (if the test is passed evaluates to NA), \code{pass} - binary result of the test either "Fail" or "Pass".

A  figure where the users \code{e(0)} values would be indicated in color and HMD values would be black:
}
\description{
Upper level function that checks the data quality. Checks if the columns are numeric, if any of the columns is missing, if there is insufficient rows, if there are missing data entries, if ages do not start with 0, and also if ages are coherent, sequential and not redundant.

Creates a plot of \code{e(0)} values from Human Mortality Database and compare it with those obtained by user.
}
\examples{
Exposures <- c(100958,466275,624134,559559,446736,370653,301862,249409,
              247473,223014,172260,149338,127242,105715,79614,53660,
              31021,16805,8000,4000,2000,1000)

Deaths <- c(8674,1592,618,411,755,1098,1100,1357,
           1335,3257,2200,4023,2167,4578,2956,4212,
           2887,2351,1500,900,500,300)

library(readr)
library(dplyr)
# single age data

}
