% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_data_graduate_time}
\alias{check_data_graduate_time}
\title{Validate input data for time graduation}
\usage{
check_data_graduate_time(data, X, Y)
}
\arguments{
\item{data}{A data frame containing the variable to be evaluated.}

\item{X}{Character value indicating name of the time variable e.g. \code{Year}, \code{Time}, \code{y}, etc.}

\item{Y}{Character value indicating name of the ariable to be evaluated, e.g. \code{Deaths}, \code{Exposures}}
}
\value{
A data frame summarizing validation results. Contains:
\itemize{
\item \code{check} – name of the validation performed
\item \code{message} – error message if the check fails, otherwise \code{NA}
\item \code{pass} – \code{"Pass"} or \code{"Fail"}
\item \code{.id} – group identifier
}
}
\description{
Performs structural and numerical validation checks for time-series data used in graduation or smoothing procedures. The function verifies that time points are numeric, unique, sequential, and evenly spaced, and that the associated values are numeric, finite, and strictly positive.
Validation is performed separately for each \code{.id} group. If \code{.id} is \code{NULL}, all observations are treated as belonging to a single group.
}
\examples{
dt <- data.frame(value = c(70.1, 70.4, 70.8),
                 time  = c(2000, 2005, 2010))

check_data_graduate_time(data = dt,
                         X = "time",
                         Y = "value")

}
