% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_data_opag}
\alias{check_data_opag}
\title{\code{check_data_opag}}
\usage{
check_data_opag(data)
}
\arguments{
\item{data}{tibble. A tibble containing population data for ODAP analysis with at minimum Age and pop columns.}
}
\value{
A data.frame with validation results containing columns: \code{check}, \code{message}, \code{pass}.
}
\description{
Upper level function that checks ODAP population data quality. Checks if the columns are numeric, if any of the columns is missing, if there is insufficient rows, if there are missing data entries, if ages do not start with 0, and also if ages are coherent, sequential and not redundant.
}
\examples{
library(tibble)
data <- tibble(
  Age = 0:100,
  pop = c(9544406, 7471790, rep(1000000, 99)),
  name = "India",
  country_code = 356,
  sex = "M",
  year = 1971
)

check_data_opag(data = data)

}
