% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetables.R
\name{check_e0}
\alias{check_e0}
\title{\code{check_e0}}
\usage{
check_e0(data_out)
}
\arguments{
\item{data_out}{a data.frame or tibble. The data.frame output of the \code{lt_flexible} function.}
}
\value{
A  figure where the users \code{e(0)} values would be indicated in color and HMD values would be black:
}
\description{
Creates a plot of \code{e(0)} values from Human Mortality Database and compare it with those obtained by user.
}
\examples{
library(readr)
library(dplyr)
# single age data
data_in <- read_csv("inst/extdata/single_hmd_spain.csv") |>
  dplyr::select(-1)

data_out <- lt_flexible(
  data_in,
  OAnew      = 100,
  age_out    = "single",
  extrapFrom = 80,
  extrapFit  = NULL,  # Default NULL, computed later
  extrapLaw  = NULL,
  radix      = 1e+05,
  SRB        = 1.05,
  a0rule     = "Andreev-Kingkade",
  axmethod   = "UN (Greville)",
  Sex = "t"
)

check_e0(data_out)

}
