% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{check_groupid}
\alias{check_groupid}
\title{check_groupid}
\usage{
check_groupid(data)
}
\arguments{
\item{data}{\code{data.frame}-like object}
}
\value{
logical TRUE if strata completely defined.
}
\description{
Checks to make sure \code{.id} column plus \code{Age} indeed completely define all strata present in the data.
}
