% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_has_numeric}
\alias{check_has_numeric}
\title{\code{check_has_numeric}}
\usage{
check_has_numeric(data)
}
\arguments{
\item{data}{tibble. A tibble with demographic data.}
}
\value{
A data.frame with validation results containing columns: \code{check}, \code{message}, \code{pass}.
}
\description{
Check if the data has at least one numeric column besides Age.
}
