% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_diagnostics.R
\name{check_heaping_general}
\alias{check_heaping_general}
\title{\code{check_heaping_general}}
\usage{
check_heaping_general(data, y)
}
\arguments{
\item{data}{data.frame. User file from the read_data command with the minimum data on \code{Exposures}, \code{Death} and \code{Age}. Data can be both in 5 and 1 year age intervals}

\item{y}{character.Variable name for which the heaping should be checked \code{Deaths} or \code{Exposures}.}
}
\value{
A data.frame with 3 columns \code{method} - the method used for age heaping evaluation, \code{result} - the resulting heaping measure, and \code{.id} - user specified groups identifier (if missing is set to "all")
}
\description{
Check the age heaping for 5 or 1 year data.
}
\examples{
\dontrun{
check_heaping_general(
    data = data,
    y = "Exposures")
}

}
