% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_diagnostics.R
\name{check_heapping_general}
\alias{check_heapping_general}
\title{check_heapping_general}
\usage{
check_heapping_general(data, y)
}
\arguments{
\item{data}{data.frame. User file from the read_data command with the minimum data on Exposures, Death and Age. Data ca be both in 5 and 1 year age intervals}

\item{y}{character.Variable name for which the hipping should be checked \code{Deaths} or \code{Exposures}.}
}
\value{
A data.frame with 2 columns \code{method} - the method used for age hipping evaluation and \code{result} - the resulting hipping measure
}
\description{
Check the age heaping for 5 or 1 year data.
}
\examples{
\dontrun{
check_heapping_general(
    data = data,
    y = "Exposures")
}

}
