% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_missing_cols_generic}
\alias{check_missing_cols_generic}
\title{\code{check_missing_cols_generic}}
\usage{
check_missing_cols_generic(data)
}
\arguments{
\item{data}{tibble. A tibble with at minimum an Age column.}
}
\value{
A data.frame with validation results containing columns: \code{check}, \code{message}, \code{pass}.
}
\description{
Check if Age column exists. For generic validation that doesn't require specific value columns.
}
