% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_missing_cols_opag}
\alias{check_missing_cols_opag}
\title{\code{check_missing_cols_opag}}
\usage{
check_missing_cols_opag(data)
}
\arguments{
\item{data}{tibble. A tibble containing population data for ODAP analysis.}
}
\value{
A data.frame with 3 columns: \code{check} - the name of the test, \code{message} - The error message with corresponding information generated if the test is failed (if the test is passed evaluates to NA), \code{pass} - binary result of the test either "Fail" or "Pass".
}
\description{
Check if any of the crucial columns are missing from ODAP population data: (\code{Age}, \code{pop}).
}
\examples{
library(tibble)
data <- tibble(
  Age = 0:100,
  pop = c(9544406, 7471790, rep(1000000, 99)),
  name = "India",
  country_code = 356,
  sex = "M",
  year = 1971
)

check_missing_cols_opag(data = data)

}
