% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_nlx_if_present}
\alias{check_nlx_if_present}
\title{\code{check_nlx_if_present}}
\usage{
check_nlx_if_present(data)
}
\arguments{
\item{data}{tibble. A tibble containing population data for ODAP analysis.}
}
\value{
A data.frame with 3 columns: \code{check} - the name of the test, \code{message} - The error message with corresponding information generated if the test is failed (if the test is passed evaluates to NA), \code{pass} - binary result of the test either "Fail" or "Pass".
}
\description{
Check if nLx column is present and valid (numeric, no NAs).
}
\examples{
library(tibble)
data <- tibble(
  Age = 0:100,
  pop = c(9544406, 7471790, rep(1000000, 99)),
  nLx = runif(101, 50000, 100000)
)

check_nlx_if_present(data = data)

}
