% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_upper}
\alias{check_upper}
\title{Validate Input Data for Demographic Modules}
\usage{
check_upper(
  data,
  module = c("generic", "opag", "lifetable", "heaping", "graduate_time"),
  X = NULL,
  Y = NULL
)
}
\arguments{
\item{data}{A data frame containing the input data to be checked.}

\item{module}{Character string specifying which module's validation
routine should be used. Options are:
\describe{
\item{generic}{General data validation.}
\item{opag}{Old-age population analysis data validation.}
\item{lifetable}{Life table data validation.}
\item{heaping}{Age heaping diagnostics data validation.}
\item{graduate_time}{Time graduation data validation.}
}}

\item{X}{Optional character string indicating the name of the
independent variable (e.g., age or time) used by some modules.}

\item{Y}{Optional character string indicating the name of the
dependent variable used by some modules.}
}
\value{
The result of the selected validation function. Typically this
is a cleaned or validated version of the input data, or an object
containing diagnostic messages.
}
\description{
\code{check_upper()} is a wrapper function that dispatches input data
validation to module-specific checking functions. Depending on the
selected module, the appropriate validation routine is applied.
}
\details{
This function acts as a dispatcher that selects the appropriate
data validation routine depending on the module specified in
\code{module}. Each module has its own structure and required variables,
so dedicated checking functions are used.
}
\examples{
library(readr)
library(dplyr)
fpath       <- system.file("extdata", "single_hmd_spain.csv.gz", package = "ODAPbackend")
data_in     <- read_csv(fpath, col_select = c(-1), show_col_types = FALSE)
data_in     <- data_in[data_in$`.id` == 1, ]
data_in <- data_in \%>\% 
  mutate(nMx = Deaths/ Exposures)
opag <- data_in \%>\% 
 rename(pop = Exposures)

check_upper(data = data_in, module = c("generic"))
check_upper(data = opag, module = c("opag"))
check_upper(data = data_in, module = c("heaping"), X = "Deaths")
check_upper(data = data_in, module = c("graduate_time"), X = "Year", Y = "Exposures")

}
