% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{create_groupid}
\alias{create_groupid}
\title{create_groupid}
\usage{
create_groupid(data, keys)
}
\arguments{
\item{data}{a \code{data.frame} or \code{tibble}}

\item{keys}{character vector of columns definining strata}
}
\description{
Checks to make sure that key variables plus age result in one row per unique stratum.
}
