% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{extension_check}
\alias{extension_check}
\title{extension_check}
\usage{
extension_check(user_file)
}
\arguments{
\item{user_file}{character. File name with corresponding extension e.g. Should be one of ".csv",".xlsx",".xls",".tsv"}
}
\value{
A character string with the corresponding file extension.
}
\description{
Check the provided files extension.
}
\examples{
\dontrun{
extension_check(
    user_file = "data.csv")
}

}
