% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate_data_first_go.R
\name{graduate_auto}
\alias{graduate_auto}
\title{graduate_auto}
\usage{
graduate_auto(
  data_in,
  age_out = "single",
  variable = "Pop",
  constrain_infants = TRUE
)
}
\arguments{
\item{data_in}{tibble. A tibble with two columns - \code{Pop} - population counts and \code{Age} provided in single age intervals, 5 year age intervals or abridged format e.g. with ages 0, 1-4, 5-9 etc}

\item{age_out}{character. A scalar with the desired age grouping output. 3 possible options - \code{single} for single ages, \code{5-year} - for 5-year age groups, and \code{abridged} - for abridged data. NOTE, \code{abridged} will not work if the data_in is in 5-year groups, since we do not know the initial distributions in the abridged ages.}

\item{variable}{character. A scalar with the variable name which is to be graduated. For example \code{Pop} or \code{Death}}

\item{constrain_infants}{logical. A scalar indicating weather the infant proportions should be constrained or left as is.}
}
\value{
A tibble with 2 columns - ypur chosen \code{variable} with graduated and smoothed counts and \code{Age}

data_out. A tibble with two numeric columns - smoothed counts for the chosen variable and \code{Age} - chosen age grouping
}
\description{
Smooths the counts using the moving averages. The methods are adoзted from the "Method protocol for the evaluation of census population data by age and sex"
}
