% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetables.R
\name{lt_flexible}
\alias{lt_flexible}
\title{lt_flexible}
\usage{
lt_flexible(
  Deaths = Deaths,
  Exposures = Exposures,
  Age = Age,
  OAnew = 100,
  age_out = "single",
  extrapFrom = 80,
  extrapFit = Age[Age >= 60],
  extrapLaw = NULL,
  radix = 1e+05,
  SRB = 1.05,
  a0rule = "ak",
  axmethod = "un",
  Sex = "m"
)
}
\arguments{
\item{Deaths}{numeric vector. Death counts in one, five or ten years age groups.}

\item{Exposures}{numeric vector. Population counts in one, five or ten years age groups.}

\item{Age}{numeric vector. Age in integer years at the beginning of the interval. e.g. interval \code{"5-10"}, should be coded as 5}

\item{OAnew}{integer. Desired open age group (5-year ages only). Default \code{100}. If higher then rates are extrapolated.}

\item{age_out}{character. Indicates weather single of abridged lif table output is desired. takes values \code{"single"}, \code{"abridged"}. Defaults to "single"}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages > =60.}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include \code{"Kannisto"}, \code{"Kannisto_Makeham"}, \code{"Makeham"}, \code{"Gompertz"}, \code{"GGompertz"}, \code{"Beard"}, \code{"Beard_Makeham"}, \code{"Quadratic"}. Default \code{"Kannisto"} if the highest age is at least 90, otherwise \code{"makeham"}.}

\item{radix}{numeric. Lifetable radix, \code{l0}. Default \code{100000}.}

\item{SRB}{numeric. the sex ratio at birth (boys / girls), default \code{1.05}}

\item{a0rule}{character. Either \code{"ak"} (default) or \code{"cd"}.}

\item{axmethod}{character. Either \code{"pas"} or \code{"un"}.}

\item{Sex}{character. Either \code{"m"} for males, \code{"f"} for females (default).}
}
\value{
A single or abridged life table iof data.frame format with corresponding columns:
Age integer. Lower bound of abridged age class,
AgeInt integer. Age class widths.
nMx numeric. Age-specific central death rates.
nAx numeric. Average time spent in interval by those deceased in interval.
nqx numeric. Age-specific conditional death probabilities.
lx numeric. Lifetable survivorship
ndx numeric. Lifetable deaths distribution.
nLx numeric. Lifetable exposure.
Sx numeric. Survivor ratios in uniform 5-year age groups.
Tx numeric. Lifetable total years left to live above age x.
ex numeric. Age-specific remaining life expectancy.
}
\description{
Calculate an abridged-age or a single-age lifetable.
}
\examples{
\dontrun{
Exposures <- c(100958,466275,624134,559559,446736,370653,301862,249409,
247473,223014,172260,149338,127242,105715,79614,53660,
31021,16805,8000,4000,2000,1000)

Deaths <- c(8674,1592,618,411,755,1098,1100,1357,
            1335,3257,2200,4023,2167,4578,2956,4212,
            2887,2351,1500,900,500,300)

Age = c(0, 1, seq(5, 100, by = 5))
data_out <- 
  lt_flexible(Deaths    = Deaths, 
              Exposures = Exposures,
              Age       = Age,
              OAnew     = 100,
              age_out   = "single",  
              extrapFrom = 80,
              extrapFit = Age[Age >= 60], 
              radix     = 1e+05,
              extrapLaw = NULL,
              SRB       = 1.05,
              a0rule    = "ak",
              axmethod  = "un",
              Sex       = "m")
}
}
