% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetables.R
\name{lt_flexible_chunk}
\alias{lt_flexible_chunk}
\title{\code{lt_flexible_chunk}}
\usage{
lt_flexible_chunk(
  data_in,
  Sex,
  OAnew = 100,
  age_out = "single",
  extrapFrom = 80,
  extrapFit = NULL,
  extrapLaw = NULL,
  radix = 1e+05,
  SRB = 1.05,
  a0rule = "Andreev-Kingkade",
  axmethod = "UN (Greville)"
)
}
\arguments{
\item{data_in}{a \code{data.frame} or \code{tibble}. 3 numeric columns \code{Age}, \code{Deaths}, and \code{Exposures}, or \code{nMx} or any of the following lifetable functions: \code{nqx}, \code{npx}, \code{ndx}, \code{lx}. Can optionally have columns \code{Sex} and \code{.id} in which case the table will be calculated for each level in these columns. Can also have an additional column.}

\item{Sex}{character. Either \code{m} for males, \code{f} for females, or \code{t} for total (default).}

\item{OAnew}{integer. Desired open age group (5-year ages only). Default \code{100}. If higher then rates are extrapolated.}

\item{age_out}{character. Indicates whether single or abridged lifetable output is desired. Takes values \code{single}, or \code{abridged}. Defaults to \code{single}.}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality. Defaults to \code{NULL}.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages >= 60.}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include \code{Kannisto}, \code{Kannisto_Makeham}, \code{Makeham}, \code{Gompertz}, \code{GGompertz}, \code{Beard}, \code{Beard_Makeham}, \code{Quadratic}. Defaults to \code{Kannisto} if the highest age is at least 90, otherwise to \code{Makeham}.}

\item{radix}{numeric. Lifetable radix, \code{l(0)}. Default \code{100000}.}

\item{SRB}{numeric. the sex ratio at birth (boys/girls). Defaults to \code{1.05}.}

\item{a0rule}{character. An a(0) calculation rule. Either \code{ak} (default) or \code{cd}.}

\item{axmethod}{character. A method used for a(0) calculation. Either \code{pas} or \code{un}.}
}
\value{
A list with two elements: A single or abridged life table of data.frame format with corresponding columns:
Age integer. Lower bound of abridged age class,
AgeInt integer. Age class widths.
nMx numeric. Age-specific central death rates.
nAx numeric. Average time spent in the interval by those deceased.
nqx numeric. Age-specific conditional death probabilities.
lx numeric. Lifetable survivorship.
ndx numeric. Lifetable deaths distribution.
nLx numeric. Lifetable exposure.
Sx numeric. Survivor ratios in uniform 5-year age groups.
Tx numeric. Lifetable total years left to live above age x.
ex numeric. Age-specific remaining life expectancy.
Sex character. Sex.
}
\description{
Calculate an abridged or a single-age lifetable.
}
\examples{
library(readr)
library(dplyr)
# single age data
fpath <- system.file("extdata", 
"single_hmd_spain.csv", 
package = "ODAPbackend")
data_in <- read_csv(fpath) |>
  dplyr::select(-1) |>
  filter(.id == 1)|> 
  # This step happens in lt_flexible...
  mutate(Mx_emp = Deaths / Exposures)

data_out <-
  lt_flexible_chunk(data_in,
                    OAnew     = 100,
                   age_out   = "abridged",
                   extrapFrom = 80,
                   # extrapFit = Age[Age >= 60],
                   radix     = 1e+05,
                    extrapLaw = NULL,
                    SRB       = 1.05,
                   a0rule    = "ak",
                   axmethod  = "un",
                   Sex       = "f")

data_out$data_out


}
