% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetables.R
\name{lt_plot}
\alias{lt_plot}
\title{lt_plot}
\usage{
lt_plot(data_in, data_out, extrapFrom = extrapFrom)
}
\arguments{
\item{data_in}{a \code{data.frame} or \code{tibble} with columns \code{Age}, \code{Deaths}, and \code{Exposures} and \code{.id}}

\item{data_out}{\code{tibble} as produced by \code{lt_flexible()}}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}
}
\description{
Plot wrapper, created lifetable plot list previously returned by \code{lt_flexible()}
}
\details{
This function should be run after \code{lt_flexible()}, so that you can pass both \code{data_in} and \code{data_out}. There is no fallback at this time to generate \code{data_out} on the fly if missing. We need to pass \code{extrapFrom} at this time indicate the jump-off in the plot. In the future this may be detected or passed in another way.
}
