% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetables.R
\name{lt_summary_chunk}
\alias{lt_summary_chunk}
\title{\code{lt_summary_chunk}}
\usage{
lt_summary_chunk(data_out, i18n = NULL)
}
\arguments{
\item{data_out}{a data.frame or tibble. The data.frame output of the \code{lt_flexible} function.}

\item{i18n}{An optional i18n object for translation.}
}
\value{
A  data.frame containing the information on the following useful lifetable statistics:
e0 - life expectancy at birth
e65 - life expectancy at age 65
\code{S[1]} - The standard deviation in age at death from birth
\code{S[11]} - The standard deviation in age at death from age 10
IQR1 - interquartile range survivorship age from a lifetable
IQR2 - interquartile range survivorship age from a lifetable
IQR3 - interquartile range survivorship age from a lifetable
mod_age - modal age at death
q15_45 - probability that the person ages 54 will die at age 60
}
\description{
Creates a table of useful summary statistics from the lifetable.
}
\examples{
library(readr)
library(dplyr)
# single age data
fpath <- system.file("extdata", 
"single_hmd_spain.csv.gz", 
package = "ODAPbackend")
data_in <- read_csv(fpath) |>
  dplyr::select(-1) |>
  filter(.id == 1)

data_out <- lt_flexible(
  data_in,
  OAnew      = 100,
  age_out    = "single",
  extrapFrom = 80,
  extrapFit  = NULL,  # Default NULL, computed later
  extrapLaw  = NULL,
  radix      = 1e+05,
  SRB        = 1.05,
  a0rule     = "Andreev-Kingkade",
  axmethod   = "UN (Greville)"
)

lt_summary_chunk(data_out$data_out)

}
