% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_initial}
\alias{plot_initial}
\title{plot_initial}
\usage{
plot_initial(
  data,
  plot_exposures = TRUE,
  plot_deaths = TRUE,
  plot_rates = TRUE
)
}
\arguments{
\item{data}{tibble. Empirical data downloaded  with the \code{read_data} function}

\item{plot_exposures}{logical indicates weather the population pyramid should be plotted, defaults to TRUE}

\item{plot_deaths}{logical indicates weather the death pyramid should be plotted, defaults to TRUE}

\item{plot_rates}{logical indicates weather the empirical \code{M(x)} should be plotted, defaults to TRUE}
}
\value{
A named list with 3 elements: \code{Exposures} - population pyramid, \code{Deaths} - death pyramid and \verb{Empirical Mx} - log 10 transformed empirical \code{M(x)} value
}
\description{
Plots a line graph of the log10 transformed empirical mortality rate \code{M(x)}, population pyramid and death pyramid if the data contains information on two sex.
}
\examples{
\dontrun{
plot_initial(data = mutate(data, sex = "Female"), plot_exposures = TRUE, plot_deaths = TRUE, plot_rates = TRUE)
}
}
