% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_input_rates}
\alias{plot_input_rates}
\title{plot_input_rates}
\usage{
plot_input_rates(data)
}
\arguments{
\item{data}{tibble. Empirical data downloaded  with the \code{read_data} function}
}
\value{
A linechart of log 10 scaled empirical \code{M(x)} values
}
\description{
Plots a line graph of the log10 transformed empirical mortality rate (Mx).
}
\examples{
\dontrun{
plot_input_rates(data = mutate(data, sex = "Female"))
}

}
