% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lifetable}
\alias{plot_lifetable}
\title{plot_lifetable}
\usage{
plot_lifetable(data_out)
}
\arguments{
\item{data_out}{tibble. Modeled lifetable generated by the lt_flexible function.}
}
\value{
A ggplot line graph with facets showing the following variables: ex, lx, nAx, ndx, nLx, nMx, nqx, Sx, Tx. The dotted lines for ndx and lx correspond to .25, 0.5, and 0.75 quartiles of the distribution and the solid line is e0. Graphs created for each group identified by .id argument
}
\description{
Creates a list of plots of lifetable functions. Can work with groups identified by .id argument
}
\examples{
\dontrun{
library(tibble)
Exposures <- c(
  100958, 466275, 624134, 559559, 446736, 370653, 301862, 249409,
  247473, 223014, 172260, 149338, 127242, 105715, 79614, 53660,
  31021, 16805, 8000, 4000, 2000, 1000
)

Deaths <- c(
  8674, 1592, 618, 411, 755, 1098, 1100, 1357,
  1335, 3257, 2200, 4023, 2167, 4578, 2956, 4212,
  2887, 2351, 1500, 900, 500, 300
)

Age <- c(0, 1, seq(5, 100, by = 5))
data_in <- tibble(Age, Deaths, Exposures)
data_out <-
  lt_flexible(data_in,
    OAnew = 100,
    age_out = "single",
    extrapFrom = 80,
    extrapFit = Age[Age >= 60],
    radix = 1e+05,
    extrapLaw = NULL,
    SRB = 1.05,
    a0rule = "ak",
    axmethod = "un",
    Sex = "m"
  )
plots <- plot_lifetable(data_out = data_out)
print(plots$nMx)
print(plots$lx)
print(plots$ndx)
}
}
