% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{plot_smooth_compare}
\alias{plot_smooth_compare}
\title{\code{plot_smooth_compare}}
\usage{
plot_smooth_compare(data_in, data_out, variable, i18n = NULL)
}
\arguments{
\item{data_in}{tibble or data.frame. A tibble with two numeric columns - population or death counts - provided in single age intervals, 5-year age intervals, or abridged age format e.g. with ages 0, 1, 5 etc, and corresponding Age \code{column}.}

\item{data_out}{A tibble with two numeric columns - smoothed counts for the chosen variable as produced by the \code{smooth_flexible}, or \code{smooth_flexible_chunk}, and \code{Age} - corresponding Age column.}

\item{variable}{character. A scalar with the \code{variable} name which is to be plotted. The list of possible options includes \code{Pop}, \code{Population}, \code{Exp}, \code{Exposures} or \code{Deaths}.}

\item{i18n}{An optional i18n object for translation.}
}
\value{
list. A named list with 3 elements: \code{figure} - a plot of original versus adjusted data, \code{data_adjusted} - and \code{data_original}.
}
\description{
Plots the original counts and the smoothed counts over the chosen age range for ease of comparison.
}
\examples{
library(readr)
library(dplyr)
# single age data
fpath <- system.file("extdata", 
"abridged_hmd_spain.csv.gz", 
 package = "ODAPbackend")
data_in <- read_csv(fpath, show_col_types = FALSE) |>
  select(-1) |>
  filter(.id == 1)

data_out <- smooth_flexible_chunk(
  data_in,
  variable     = "Exposures",
  rough_method = "auto",
  fine_method  = "auto",
  constrain_infants = TRUE,
  age_out      = "single",
  u5m          = NULL,
  Sex          = "t"
)

results <- plot_smooth_compare(data_in,
                               data_out$data,
                               variable = "Exposures")

# smoothed data
results$data_adjusted
# original data
results$data_original
# figure
results$figure

}
