% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate_data_first_go.R
\name{procedure_for_5_y_ages}
\alias{procedure_for_5_y_ages}
\title{procedure_for_5_y_ages}
\usage{
procedure_for_5_y_ages(dat_5)
}
\arguments{
\item{dat_5}{tibble. A tibble with two columns - \code{Pop} - population counts by 5 year of age and corresponding \code{Age} provided in 5 year age intervals}
}
\value{
A named vector with graduated and smoothed population counts
}
\description{
Implements the method protocol procedure for data with 5-year age groups
}
\examples{
\dontrun{
dat <- tibble(Pop = pop1m_ind,
Age = 0:100)

gradute_data(dat = dat)
}

}
