% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_data}
\alias{read_data}
\title{read_data}
\usage{
read_data(user_file, skip = 0)
}
\arguments{
\item{user_file}{character. File name with corresponding extension e.g. \code{data.csv}.}

\item{skip}{numeric. Number of rows to skip before reading the data. Defaults to \code{0}.}
}
\value{
A tibble with with 5 numeric columns: Deaths, Exposures, Age, AgeInt, Mx_emp.
}
\description{
Read the with supported file extension into the program. The file should contain at minimum 3 columns "Deaths", "Exposures","Age" named this way or positioned in a corresponding order.
}
\examples{
\dontrun{
read_data(
    user_file = "data.csv",
    skip = 2)
}
}
