% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing1d.R
\name{smooth1d_chunk}
\alias{smooth1d_chunk}
\title{smooth1d_chunk}
\usage{
smooth1d_chunk(
  data_in.,
  method = c("supsmu", "lowess", "loess", "cubicspline", "gam-tp", "gam-ps")[1],
  smoothing_par = 1,
  xout = data_in.[["x"]]
)
}
\arguments{
\item{data_in.}{\code{data.frame} with x and y coordinates to fit to. Optionally with \code{weights}. Should refer to a single subset of data.}

\item{method}{character. Smoothing method desired. options \code{"supsmu"} (default),\code{"lowess"},\code{"loess"},\code{"cubicspline"},\code{"gam-tp"},\code{"gam-ps"}}

\item{smoothing_par}{smoothing parameter, interpretation varies by method, but higher always means smoother.Default 1}

\item{xout}{numeric vector of coordinates to predict for. Defaults to original unique coordinates.}

\item{xname}{name of variable containing x coordinate, default \code{x}}

\item{yname}{name of variable containing y coordinate, default \code{y}}
}
\description{
Smooth a univariate time series, optionally using \code{weights.} Choose between the super-smoother (\code{"supsmu"}) method, loess  (\code{"lowess"} or \code{"loess"}) , smoothing splines (\code{"cubicsplines"}), thin-plate splines (\code{"gam-tp"}), or p-splines  (\code{"gam-ps"}). Input data may have multiple observations per x coordinate. Output can be for arbitrary x-coordinates (\code{xout}).
}
\details{
\code{"supsmu"} method takes a \code{smoothing_par} between 0 and 10. \code{"lowess"} and
}
