% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{smooth_flexible}
\alias{smooth_flexible}
\title{\code{smooth_flexible}}
\usage{
smooth_flexible(
  data_in,
  variable = "Deaths",
  age_out = c("single", "abridged", "5-year"),
  fine_method = c("auto", "none", "sprague", "beers(ord)", "beers(mod)", "grabill",
    "pclm", "mono", "uniform"),
  rough_method = c("auto", "none", "Carrier-Farrag", "KKN", "Arriaga", "United Nations",
    "Strong", "Zigzag"),
  u5m = NULL,
  constrain_infants = TRUE,
  Sex = "t",
  by_args = NULL,
  i18n = NULL
)
}
\arguments{
\item{data_in}{tibble or data.frame. A tibble with two numeric columns - population or death counts with supported names: \code{Pop}, \code{Population}, \code{Exp}, \code{Exposures} or \code{Deaths}, and corresponding numeric \code{Age} - provided in single age intervals, 5-year age intervals, or abridged age format e.g. with ages 0, 1, 5, etc.}

\item{variable}{character. A scalar with the variable name which is to be graduated. The list of possible options includes \code{Pop}, \code{Population}, \code{Exp}, \code{Exposures} or \code{Deaths}.}

\item{age_out}{character. The desired age structure of the output file. Possible options include \code{single} - for single years, \code{5-year} - for 5-year data, and \code{abridged} - for abridged data, e.g. 0, 1, 5, etc.}

\item{fine_method}{the \code{method} argument of \code{graduate()} function from \code{DemoTools} that graduates grouped data. Possible options include  \code{auto}, \code{none}, \code{sprague}, \code{beers(ord)}, \code{beers(mod)}, \code{grabill}, \code{pclm}, \code{mono}, \code{uniform}.}

\item{rough_method}{the \code{method} argument of \code{smooth_age_5()} function from \code{DemoTools} that smoothes populations in 5-year age groups using various methods. Possible options include \code{auto}, \code{none}, \code{Carrier-Farrag}, \code{KKN}, \code{Arriaga}, \verb{United Nations}, \code{Strong}, \code{Zigzag}.}

\item{u5m}{numeric. Under-five mortality rate. Defaults to \code{NULL}.}

\item{constrain_infants}{logical, if age 0 is a separate age class, shall we constrain its proportion within the age group 0-5 in the output? Default to \code{TRUE}.}

\item{Sex}{character. Either \code{m} for males, \code{f} for females, or \code{t} for total (default). Please note that in case this parameter is not explicitly provided, the function will scan the data for the column with the corresponding name and use its levels automatically.}

\item{by_args}{character. A vector of columns should be also included in the output. These columns are usually ones that are used for \code{.id} construction. Defaults to \code{NULL}. It is important to not include \code{Sex} in this vector.}

\item{i18n}{An optional i18n object for translation.}
}
\value{
data_out. A list of lists. For each separate group specified by the \code{.id} column, the function will generate a list that contains 2 lists: data and figures. Data tibble with two numeric columns - smoothed counts for the chosen \code{variable} and \code{Age} - chosen age grouping and one character column indicating \code{.id} groups. May optionally include the columns identified in \code{by_args}. And two figures data.frames - \code{data_adjusted} and \code{data_original} corresponding to the data.frames with pre and post-old age mortality adjustment and a \code{ggplot2} figure visualizing the corresponding differences.
}
\description{
This is a wrapper around \code{smooth_flexible_chunk} that allows us to work with \code{.id} variable and create the output for corresponding groups. Smoothes population or death counts using various methods from the \code{DemoTools} package and paragraph five from "Method protocol for evaluating census population data by age and sex".
}
\examples{
library(readr)
library(dplyr)
# single age data
fpath <- system.file("extdata", 
"single_hmd_spain.csv.gz", 
package = "ODAPbackend")
data_in <- read_csv(fpath) |>
  select(-1)
ex1 <- smooth_flexible(
 data_in,
 variable     = "Exposures",
 rough_method = "auto",
 fine_method  = "auto",
 age_out      = "abridged",
 u5m          = NULL,
 Sex          = "t",
 constrain_infants = TRUE,
 by_args      = "Year"
)

# data
ex1$data_out
# arguments
ex1$arguments
# figures
ex1$figures$`1_2012`$figure
ex1$figures$`2_2012`$figure
ex1$figures$`3_2012`$figure

}
