% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{smooth_flexible_chunk}
\alias{smooth_flexible_chunk}
\title{\code{smooth_flexible_chunk}}
\usage{
smooth_flexible_chunk(
  data_in,
  variable = "Deaths",
  age_out = c("single", "abridged", "5-year"),
  fine_method = c("auto", "none", "sprague", "beers(ord)", "beers(mod)", "grabill",
    "pclm", "mono", "uniform"),
  rough_method = c("auto", "none", "Carrier-Farrag", "KKN", "Arriaga", "United Nations",
    "Strong", "Zigzag"),
  u5m = NULL,
  Sex = c("t", "f", "m"),
  constrain_infants = TRUE,
  i18n = NULL
)
}
\arguments{
\item{data_in}{tibble or data.frame. A tibble with two numeric columns - population or death counts with supported names: \code{Pop}, \code{Population}, \code{Exp}, \code{Exposures} or \code{Deaths}, and corresponding \code{Age} column - provided in single age intervals, 5-year age intervals, or abridged age format e.g. with ages 0, 1, 5 etc.}

\item{variable}{character. A scalar with the \code{variable} name which is to be graduated. The list of possible options includes \code{Pop}, \code{Population}, \code{Exp}, \code{Exposures} or \code{Deaths}.}

\item{age_out}{character. The desired age structure of the output file. Possible options include \code{single} - for single years, \code{5-year} - for 5-year data, and \code{abridged} - for abridged data, e.g. 0, 1, 5, etc.}

\item{fine_method}{the \code{method} argument of \code{graduate()} function from \code{DemoTools} that graduates grouped data. Possible options include  \code{auto}, \code{none}, \code{sprague}, \code{beers(ord)}, \code{beers(mod)}, \code{grabill}, \code{pclm}, \code{mono}, \code{uniform}.}

\item{rough_method}{the \code{method} argument of \code{smooth_age_5()} function from \code{DemoTools} that smoothes populations in 5-year age groups using various methods. Possible options include \code{auto}, \code{none}, \code{Carrier-Farrag}, \code{KKN}, \code{Arriaga}, \verb{United Nations}, \code{Strong}, \code{Zigzag}.}

\item{u5m}{numeric. Under-five mortality rate. Defaults to \code{NULL}.}

\item{Sex}{character. Either \code{m} for males, \code{f} for females, or \code{t} for total (default). Please note that in case this parameter is not explicitly provided, the function will scan the data for the column with the corresponding name and use its levels automatically.}

\item{constrain_infants}{logical. If age 0 is a separate age class, shall we constrain its proportion within the age group 0-5 in the output? Default to \code{TRUE}.}

\item{i18n}{An optional i18n object for translation.}
}
\value{
data_out. A list with 3 elements: - \code{data} - tibble with two numeric columns - smoothed counts for the chosen \code{variable}, \code{Age} - chosen age grouping, and \code{Sex}, \code{figure} - a figure with corresponding data illustrating pre and post-adjustment data, \code{arguments} - a list of arguments used in the function.
}
\description{
Smoothes population or death counts using a variety of methods from the \code{DemoTools} package and paragraph five of the "Method protocol for the evaluation of census population data by age and sex".
}
\examples{
library(readr)
library(dplyr)
# single age data
fpath <- system.file("extdata", 
"abridged_hmd_spain.csv.gz", 
 package = "ODAPbackend")
data_in <- read_csv(fpath) |>
  select(-1) |>
  filter(.id == 1)

ex1 <- smooth_flexible_chunk(
  data_in,
  variable     = "Exposures",
  rough_method = "auto",
  fine_method  = "auto",
  constrain_infants = TRUE,
  age_out = "single",
  u5m     = NULL,
  Sex     = "t")

ex1$data$Exposures
ex1$figure$figure
ex1$arguments
}
