% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate_data_first_go.R
\name{graduate_auto}
\alias{graduate_auto}
\title{graduate_auto}
\usage{
graduate_auto(dat)
}
\arguments{
\item{dat}{tibble. A tibble with two columns - \code{Pop} - population counts and \code{Age} provided in single age intervals, 5 year age intervals or abridged format e.g. with ages 0, 1-4, 5-9 etc}
}
\value{
A named vector with graduated and smoothed population counts
}
\description{
Smooths the counts using the moving averages. The methods are adoзted from the "Method protocol for the evaluation of census population data by age and sex"
}
\examples{
\dontrun{
dat <- tibble(Pop = pop1m_ind,
Age = 0:100)

graduate_auto(dat = dat)
}


}
