% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduate_data_first_go.R
\name{graduate_auto_5}
\alias{graduate_auto_5}
\title{graduate_auto_5}
\usage{
graduate_auto_5(dat_5)
}
\arguments{
\item{dat_5}{tibble. A tibble with two columns - \code{Pop} - population counts in 5-year age groups and corresponding \code{Age} column (lower bound of age group)}
}
\value{
A named vector with graduated and smoothed population counts
}
\description{
Implements the method protocol procedure for data with 5-year age groups
}
