% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetables.R
\name{lt_flexible}
\alias{lt_flexible}
\title{\code{lt_flexible}}
\usage{
lt_flexible(
  data_in,
  OAnew = 100,
  age_out = "single",
  extrapFrom = 80,
  extrapFit = NULL,
  extrapLaw = NULL,
  radix = 1e+05,
  SRB = 1.05,
  a0rule = "Andreev-Kingkade",
  axmethod = "UN (Greville)",
  Sex = "t",
  by_args = NULL
)
}
\arguments{
\item{data_in}{data.frame or tibble. Should contain numeric columns \code{Age}, \code{Deaths}, and \code{Exposures}, or \code{nMx} or any of the following life table functions: \code{nqx}, \code{ndx}, \code{lx}. Can have a \code{.id} column in which case the results will be generated for each group as specified by this column.}

\item{OAnew}{integer. Desired open age group (5-year ages only). Default to \code{100}. If higher, then rates will be extrapolated.}

\item{age_out}{character. Indicates whether single or abridged life table output is desired. Takes values \code{single}, or \code{abridged}. Defaults to \code{single}.}

\item{extrapFrom}{integer. Age from which to impute extrapolated mortality.}

\item{extrapFit}{integer vector. Ages to include in model fitting. Defaults to all ages >= 60.}

\item{extrapLaw}{character. If extrapolating, which parametric mortality law should be invoked? Options include \code{"Kannisto"}, \code{"Kannisto_Makeham"}, \code{"Makeham"}, \code{"Gompertz"}, \code{"GGompertz"}, \code{"Beard"}, \code{"Beard_Makeham"}, \code{"Quadratic"}. Defaults to \code{Kannisto} if the highest age is at least 90, otherwise o \code{Makeham}.}

\item{radix}{numeric. Life table radix, \code{l(0)}. Defaults to \code{100000}.}

\item{SRB}{numeric. The sex ratio at birth (boys/girls). Defaults to \code{1.05}.}

\item{a0rule}{character. Rule for \code{a(x)} calculation Either \code{ak} (default) or \code{cd}.}

\item{axmethod}{character. Method used for \code{a(x)} calculation. Either \code{pas} or \code{un}.}

\item{Sex}{character. Either \code{m} for males, \code{f} for females or \code{t} for total. This variable defaults to \code{t}. If there is more than one sex in the data, then the life table will be calculated for each sex.}

\item{by_args}{character. A vector of columns should be also included in the output. These columns are usually ones that are used for \code{.id} construction. Defaults to \code{NULL}. It is important to not include \code{Sex} in this vector.}
}
\value{
A list with 3 elements: \code{data_out} - single or abridged life table of data.frame format with corresponding columns:
Age integer. Lower bound of abridged age class,
AgeInt integer. Age class widths.
nMx numeric. Age-specific central death rates.
nAx numeric. Average time spent in the interval by those deceased.
nqx numeric. Age-specific conditional death probabilities.
lx numeric. Lifetable survivorship
ndx numeric. Lifetable deaths distribution.
nLx numeric. Lifetable exposure.
Sx numeric. Survivor ratios in uniform 5-year age groups.
Tx numeric. Lifetable total years left to live above age x.
ex numeric. Age-specific remaining life expectancy.
.id character A group indicator for which the results will be generated. In case of missing the .id column will return \code{all}
Sex - corresponding sex
by_args - chosen additional arguments
\code{arguments} - a list of arguments used in fitting \code{lt_flixible} function, \code{arguments2} - a list of arguments used for \code{lt_flexible_chunk} fitting.
}
\description{
Wrapper for calculation of abridged-age or single-age life table.
}
\examples{
library(readr)
library(dplyr)
library(future)
# single age data
fpath <- system.file("extdata", 
"single_hmd_spain.csv.gz", 
package = "ODAPbackend")
data_in <- read_csv(fpath) |>
  dplyr::select(-1)

data_out <- lt_flexible(
  data_in,
  OAnew      = 100,
  age_out    = "single",
  extrapFrom = 80,
  extrapFit  = NULL,  # Default NULL, computed later
  extrapLaw  = NULL,
  radix      = 1e+05,
  SRB        = 1.05,
  a0rule     = "Andreev-Kingkade",
  axmethod   = "UN (Greville)"
)$data_out

}
