% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_compare_rates}
\alias{plot_compare_rates}
\title{plot_compare_rates}
\usage{
plot_compare_rates(data_in, data_out, extrapFrom)
}
\arguments{
\item{data_in}{tibble. Empirical numeric Mx value to be plotted and and numeric Age values to be plotted.}

\item{data_out}{tibble. Modelled numeric Mx value to be plotted and and numeric Age values to be plotted.}

\item{extrapFrom}{numeric. extrapFrom an age interval from which the lifetable was extrapolated.}
}
\value{
A line graph with black line corresponding to empirical Mx data and red line corresponding to modelled Mx data from the  chosen extrapFrom value.
}
\description{
Plots a line graph of the log10 transformed empirical mortality rate (Mx) data and a fitted data from a chosen age.
}
\examples{
\dontrun{
plot_compare_rates(data_in = data_in, 
                   data_out = data_out, 
                   extrapFrom = 60)
}
}
