% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_histogram}
\alias{plot_histogram}
\title{plot_histogram}
\usage{
plot_histogram(data, y)
}
\arguments{
\item{data}{tibble. Empirical data downloaded  with the \code{read_data} function}

\item{y}{character. This argument indicats weather the \code{Exposures} or \code{Deaths} should be plotted.}
}
\value{
A histogramm for either Deaths or Exposures
}
\description{
Plots a histogram of population or death, depending on the user choice.
}
\examples{
\dontrun{
plot_histogram(data = mutate(data, sex = "Female"), y = "Deaths")
}
}
