% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_initial_data}
\alias{plot_initial_data}
\title{plot_initial_data}
\usage{
plot_initial_data(
  data,
  plot_exposures = TRUE,
  plot_deaths = TRUE,
  plot_rates = TRUE
)
}
\arguments{
\item{data}{tibble. Empirical data downloaded  with the \code{read_data} function}

\item{plot_exposures}{logical. Weather the exposures should be plotted}

\item{plot_deaths}{logical. Weather the Deaths should be plotted#' @param dplot_ratesata logical Empirical data downloaded  with the \code{read_data} function}

\item{plot_rates}{logical. Weather the rates should be plotted#' @param dplot_ratesata logical Empirical data downloaded  with the \code{read_data} function}
}
\value{
A list with 3 corresponding plots for either one or two sex.
}
\description{
Plots the corresponding 3 graphics for single sex or for both sex depending on data provided by the user.
}
\examples{
\dontrun{
plot_initial_data(mutate(data, sex = "Female"), 
                  plot_exposures = TRUE, 
                  plot_deaths = TRUE, 
                  plot_rates = TRUE)
}
}
