% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_initial_single_sex}
\alias{plot_initial_single_sex}
\title{\code{plot_initial_two_sex}}
\usage{
plot_initial_single_sex(data, i18n = NULL)
}
\arguments{
\item{data}{tibble. Empirical data that was downloaded with the \code{read_data} function.}

\item{i18n}{An optional i18n object for translation.}
}
\value{
A named list with 3 elements: \code{Exposures} - population pyramid, \code{Deaths} - death pyramid and #\verb{Empirical Mx} - log 10 transformed empirical \code{M(x)} value

A named list with 3 elements: \code{Exposures} - population pyramid, \code{Deaths} - death pyramid and \verb{Empirical Mx} - log 10 transformed empirical \code{M(x)} value.
}
\description{
Plots a line graph of the log10 transformed empirical mortality rate \code{M(x)}, population #pyramid and death pyramid if the data contains information on two sex.

Plots a line graph of the log10 transformed empirical mortality rate \code{M(x)}, population histogram and death histogram if the data contains information on only one sex.
}
\examples{
\dontrun{
data1           <- data
data1$Sex       <- "Female"
data1$Exposures <- data1$Exposures
data1$Deaths    <- data1$Deaths
data <- data |> 
 full_join(data1) |> 
 mutate(Deaths = ifelse(Sex == "Female", Deaths + rpois(22, lambda = 50), Deaths))

plot_initial_two_sex(data = data)
}
Exposures <- c(100958, 466275, 624134, 559559, 446736, 370653, 301862,
               249409, 247473, 223014, 172260, 149338, 127242, 105715,
               79614,  53660,  31021,  16805,  8000,   4000,   2000,
               1000)

Deaths <- c(8674, 1592, 618,  411,  755,  1098, 1100, 1357,
            1335, 3257, 2200, 4023, 2167, 4578, 2956, 4212,
            2887, 2351, 1500, 900,  500,  300)

sex     <- c("Female")
Age     <- c(0, 1, seq(5, 100, by = 5))
data_in <- data.frame(Age, 
                      Deaths, 
                      Exposures, 
                      AgeInt = c(diff(Age), NA),
                      Sex = sex)
ex1 <- plot_initial_single_sex(data_in)

ex1$Exposures
ex1$Deaths
ex1$`Empirical Mx`

}
