% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_lifetable}
\alias{plot_lifetable}
\title{\code{plot_lifetable}}
\usage{
plot_lifetable(data_out, i18n = NULL)
}
\arguments{
\item{data_out}{tibble or data.frame. Modeled life table generated by the 'lt_flexible` function.}

\item{i18n}{An optional i18n object for translation.}
}
\value{
A ggplot2 line graph with facets showing the following variables: \code{ex}, \code{lx}, \code{nAx}, \code{ndx}, \code{nLx}, \code{nMx}, \code{nqx}, \code{Sx}, \code{Tx}. The dotted lines for \code{ndx} and \code{lx} correspond to \code{0.25}, \code{0.5}, and \code{0.75} quartiles of the distribution and the solid line is \code{e0.} Graphs will be created for each group identified by \code{.id} argument.
}
\description{
Creates a list of plots of life table functions. Can work with groups identified by \code{.id} argument.
}
\examples{
library(tibble)
Exposures <- c(
  100958, 466275, 624134, 559559, 446736, 370653, 301862, 249409,
  247473, 223014, 172260, 149338, 127242, 105715, 79614, 53660,
  31021, 16805, 8000, 4000, 2000, 1000
)

Deaths <- c(
  8674, 1592, 618, 411, 755, 1098, 1100, 1357,
  1335, 3257, 2200, 4023, 2167, 4578, 2956, 4212,
  2887, 2351, 1500, 900, 500, 300
)

Age <- c(0, 1, seq(5, 100, by = 5))
data_in <- tibble(Age, Deaths, Exposures)
data_out <-
  lt_flexible(data_in,
    OAnew = 100,
    age_out = "single",
    extrapFrom = 80,
    extrapFit = Age[Age >= 60],
    radix = 1e+05,
    extrapLaw = NULL,
    SRB = 1.05,
    a0rule = "ak",
    axmethod = "un",
    Sex = "m"
  )$data_out
result <- plot_lifetable(data_out)
result$nMx$nMx_plot
result$lx$lx_plot
result$ndx$ndx_plot
result$nqx$nqx_plot

}
