% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{plot_smooth_compare}
\alias{plot_smooth_compare}
\title{plot_smooth_compare compares the original and smoothed counts over the chosen age ranges}
\usage{
plot_smooth_compare(data_in, data_out, variable)
}
\arguments{
\item{data_in}{tibble. A tibble with two numeric columns - population or death counts - provided in single age intervals, 5-year age intervals, or abridged age format e.g. with ages 0, 1-4, 5-9 etc.}

\item{data_out}{A tibble with two numeric columns - smoothed counts for the chosen variable and \code{Age} - chosen age grouping. The output from the \code{smooth_flexible} function.}

\item{variable}{character. A scalar with the variable name which is to be graduated. The list of possible options include \code{Pop}, \code{Population}, \code{Exp}, \code{Exposures} or \code{Deaths}.}
}
\value{
list. A named list with 2 elements: \code{before} - plot of original input data and \code{after} - plot of smoothed data
#' @examples
This is just test settings used for light live coding.
data(pop1m_ind, package = "DemoTools")
data_in <- data.frame(Exposures = pop1m_ind,
Age = 0:100)

data_out <- smooth_flexible(
data_in,
variable     = "Exposures",
rough_method = "auto",
fine_method  ="pclm",
constrain_infants = TRUE,
age_out = "5-year",
u5m     = .1,
Sex     = "t")

plot_smooth_compare(data_in, data_out, variable = "Exposures")
}
\description{
Plots the original counts and the smoothed counts over the chocosen age ranges for ease of comparison.
}
