% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pyramid}
\alias{pyramid}
\title{pyramid}
\usage{
pyramid(data, y)
}
\arguments{
\item{data}{tibble. Empirical data downloaded  with the \code{read_data} function. Should contain the Exposures, Deaths}

\item{y}{character. This argument indicating weather the \code{Exposures} or \code{Deaths} should be plotted.}
}
\value{
A pyramid for either Deaths or Exposures
}
\description{
Generate a population pyramid from the user data. WARNING contains some tidy evaluation
}
\examples{
\dontrun{
data1           <- data
data1$Sex       <- "Female"
data1$Exposures <- data1$Exposures
data1$Deaths    <- data1$Deaths
data <- data \%>\% 
 full_join(data1) \%>\%
 mutate(Deaths = ifelse(Sex == "Female", Deaths + rpois(22, lambda = 50), Deaths))

pyramid(data = data, y = "Deaths")
}
}
