% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graduation.R
\name{smooth_flexible}
\alias{smooth_flexible}
\title{smooth_flexible wrapper aroung smooth_flexible_chunk allows to work with .id variable and create the output for groups.}
\usage{
smooth_flexible(
  data_in,
  variable = "Deaths",
  age_out = c("single", "abridged", "5-year"),
  fine_method = c("auto", "none", "sprague", "beers(ord)", "beers(mod)", "grabill",
    "pclm", "mono", "uniform"),
  rough_method = c("auto", "none", "Carrier-Farrag", "KKN", "Arriaga", "United Nations",
    "Strong", "Zigzag"),
  u5m = NULL,
  constrain_infants = TRUE,
  Sex
)
}
\arguments{
\item{data_in}{tibble. A tibble with two numeric columns - population or death counts with supported names: \code{Pop}, \code{Population}, \code{Exp}, \code{Exposures} or \code{Deaths}, and corresponding \code{Age} - provided in single age intervals, 5-year age intervals, or abridged age format e.g. with ages 0, 1-4, 5-9 etc.}

\item{variable}{character. A scalar with the variable name which is to be graduated. The list of possible options include \code{Pop}, \code{Population}, \code{Exp}, \code{Exposures} or \code{Deaths}.}

\item{age_out}{character. The desired age structure of the output file. Possible options include \code{single} - for sinle years, \code{5-year} - for 5-year data, and \code{abridged} - for abridged data, e.g 0, 1-4, 5-9, etc.}

\item{fine_method}{the \code{method} argument of \code{graduate()} function from DemoTools that graduates grouped data. Possible options include  \code{auto}, \code{none}, \code{sprague}, \code{beers(ord)}, \code{beers(mod)}, \code{grabill}, \code{pclm}, \code{mono}, \code{uniform}}

\item{rough_method}{the \code{method} argument of \code{smooth_age_5()} function from DemoTools that Smoothes populations in 5-year age groups using various methods. Possible options include \code{auto}, \code{none}, \code{Carrier-Farrag}, \code{KKN}, \code{Arriaga}, \verb{United Nations}, \code{Strong}, \code{Zigzag}.}

\item{u5m}{numeric. Under five mortality rate.}

\item{constrain_infants}{logical, if age 0 is a separate age class, shall we constrain its proportion within the age group 0-5 in the output? Default \code{TRUE}.}

\item{Sex}{character. Either \code{"m"} for males, \code{"f"} for females, or \code{"t"}. This variable should exist in the dataframe. If more than 1 sex is present, then graduation will be done for all the sexes present in the data}
}
\value{
data_out. A list of lists. For each sepate group specified by the .id column the function will generate a list that will contain 2 lists: data and figures. Data tibble with two numeric columns - smoothed counts for the chosen variable and \code{Age} - chosen age grouping and one character column indicating .id groups. And figures 2 dataframes - data_adjusted and data_orioginal correspodnign to the dataframes withpre and post old age mortality adjustment and a figure visualizing the corresponding differences.
}
\description{
Smoothes population or death counts using the variaety of methods from DemoTools packgae and "Method protocol for the evaluation of census population data by age and sex" paragraph 5.
}
\examples{
data(pop1m_ind, package = "DemoTools")
data_in <- data.frame(Exposures = pop1m_ind,
                      Age       = 0:100,
                      Sex = "f")
                      
ex1 <- smooth_flexible(
data_in, 
variable     = "Exposures",
rough_method = "auto",
fine_method  = "none", 
constrain_infants = TRUE, 
age_out = "abridged", 
u5m     = NULL)


}
